fix: errors with moving crates around
