+++ Cargo.toml
@@ -22,7 +22,6 @@ members = [
     "crates/errors",
     "crates/sp1",
     "crates/storage",
-    "crates/config",
     "crates/da",
 ]
 default-members = [
@@ -32,7 +31,6 @@ default-members = [
     "crates/groth16",
     "crates/errors",
     "crates/storage",
-    "crates/config",
 ]
 resolver = "2"
 
@@ -83,7 +81,6 @@ bincode = "1.3.3"
 sp1-zkvm = { version = "1.2.0" }
 sp1-sdk = { version = "1.2.0" }
 prism-common = { path = "crates/common" }
-prism-config = { path = "crates/config" }
 prism-storage = { path = "crates/storage" }
 prism-nova = { path = "crates/nova" }
 prism-da = { path = "crates/da" }
