+++ crates/common/src/lib.rs
@@ -1,3 +1,4 @@
 pub mod hashchain;
 pub mod operation;
 pub mod tree;
+pub mod signedcontent;
