+++ crates/prism/Cargo.toml
@@ -51,6 +51,7 @@ auto_impl = { workspace = true }
 prism-common = { workspace = true }
 prism-storage = { workspace = true }
 prism-errors = { workspace = true }
+prism-da = { workspace = true }
 prism-groth16 = { workspace = true }
 sp1-sdk = { workspace = true }
