+++ crates/prism/src/lib.rs
@@ -1,6 +1,4 @@
 pub mod cfg;
-pub mod consts;
-pub mod da;
 pub mod node_types;
 pub mod utils;
 pub mod webserver;
