+++ crates/prism/src/utils.rs
@@ -2,18 +2,11 @@ use anyhow::Result;
 use base64::{engine::general_purpose::STANDARD as engine, Engine as _};
 use bellman::groth16::{self, VerifyingKey};
 use bls12_381::{Bls12, Scalar};
-use ed25519::Signature;
 use ed25519_dalek::{Verifier, VerifyingKey as Ed25519VerifyingKey};
 use indexed_merkle_tree::tree::Proof;
+use prism_common::signedcontent::SignedContent;
 use prism_common::tree::Digest;
 use prism_errors::{GeneralError, PrismError, ProofError};
-
-#[allow(dead_code)]
-pub fn parse_json_to_proof(json_str: &str) -> Result<Proof, Box<dyn std::error::Error>> {
-    let proof: Proof = serde_json::from_str(json_str)?;
-
-    Ok(proof)
-}
 
 pub fn decode_public_key(pub_key_str: &String) -> Result<Ed25519VerifyingKey> {
     // decode the public key from base64 string to bytes
@@ -27,6 +20,35 @@ pub fn decode_public_key(pub_key_str: &String) -> Result<Ed25519VerifyingKey> {
 
     Ed25519VerifyingKey::from_bytes(&public_key_array)
         .map_err(|_| GeneralError::DecodingError("ed25519 verifying key".to_string()).into())
+}
+
+// verifies the signature of a given signable item and returns the content of the item if the signature is valid
+pub fn verify_signature<T: SignedContent>(
+    item: &T,
+    optional_public_key: Option<String>,
+) -> Result<Vec<u8>> {
+    let public_key_str = match optional_public_key {
+        Some(key) => key,
+        None => item.get_public_key()?,
+    };
+
+    let public_key = decode_public_key(&public_key_str)
+        .map_err(|_| PrismError::General(GeneralError::InvalidPublicKey))?;
+
+    let content = item.get_plaintext()?;
+    let signature = item.get_signature()?;
+
+    match public_key.verify(content.as_slice(), &signature) {
+        Ok(_) => Ok(content),
+        Err(e) => Err(GeneralError::InvalidSignature(e).into()),
+    }
+}
+
+#[allow(dead_code)]
+pub fn parse_json_to_proof(json_str: &str) -> Result<Proof, Box<dyn std::error::Error>> {
+    let proof: Proof = serde_json::from_str(json_str)?;
+
+    Ok(proof)
 }
 
 #[allow(dead_code)]
@@ -58,34 +80,6 @@ pub fn validate_epoch(
         .map_err(|e| PrismError::Proof(ProofError::VerificationError(e.to_string())))?;
 
     Ok(proof)
-}
-
-pub trait SignedContent {
-    fn get_signature(&self) -> Result<Signature>;
-    fn get_plaintext(&self) -> Result<Vec<u8>>;
-    fn get_public_key(&self) -> Result<String>;
-}
-
-// verifies the signature of a given signable item and returns the content of the item if the signature is valid
-pub fn verify_signature<T: SignedContent>(
-    item: &T,
-    optional_public_key: Option<String>,
-) -> Result<Vec<u8>> {
-    let public_key_str = match optional_public_key {
-        Some(key) => key,
-        None => item.get_public_key()?,
-    };
-
-    let public_key = decode_public_key(&public_key_str)
-        .map_err(|_| PrismError::General(GeneralError::InvalidPublicKey))?;
-
-    let content = item.get_plaintext()?;
-    let signature = item.get_signature()?;
-
-    match public_key.verify(content.as_slice(), &signature) {
-        Ok(_) => Ok(content),
-        Err(e) => Err(GeneralError::InvalidSignature(e).into()),
-    }
 }
 
 #[cfg(test)]
