+++ crates/storage/src/redis.rs
@@ -12,6 +12,7 @@ use std::{
     thread::sleep,
     time::Duration,
 };
+use serde::{Deserialize, Serialize};
 
 use prism_common::{
     hashchain::{Hashchain, HashchainEntry},
@@ -19,10 +20,23 @@ use prism_common::{
     tree::Digest,
 };
 use prism_errors::{DatabaseError, GeneralError, PrismError};
-use prism_config::RedisConfig;
 
 use crate::database::{convert_to_connection_error, Database};
 use log::debug;
+
+
+#[derive(Debug, Serialize, Deserialize, Clone)]
+pub struct RedisConfig {
+    pub connection_string: String,
+}
+
+impl Default for RedisConfig {
+    fn default() -> Self {
+        RedisConfig {
+            connection_string: "redis://127.0.0.1/".to_string(),
+        }
+    }
+}
 
 
 // there are different key prefixes for the different tables in the database
